HBASE-22444 Backport HBASE-19954 to branch-1
